Move SingleClickButton to core
SingleClickButton is a utility used in a number of modules and such should be
in core. While the app seems to built with the button in app, this wrong
structure should be fixed.

As the button is a UI element, the patch also moves the class to the package
info.nightscout.androidaps.utils.ui.
