+++ app/src/main/res/layout/actions_fragment.xml
@@ -25,7 +25,7 @@
             android:padding="10dip"
             app:columnCount="2">
 
-            <info.nightscout.androidaps.utils.SingleClickButton
+            <info.nightscout.androidaps.utils.ui.SingleClickButton
                 android:id="@+id/actions_profileswitch"
                 style="@style/ButtonSmallFontStyle"
                 android:layout_width="0dp"
@@ -40,7 +40,7 @@
                 app:layout_gravity="fill"
                 app:layout_row="0" />
 
-            <info.nightscout.androidaps.utils.SingleClickButton
+            <info.nightscout.androidaps.utils.ui.SingleClickButton
                 android:id="@+id/actions_temptarget"
                 style="@style/ButtonSmallFontStyle"
                 android:layout_width="0dp"
@@ -55,7 +55,7 @@
                 app:layout_gravity="fill"
                 app:layout_row="0" />
 
-            <info.nightscout.androidaps.utils.SingleClickButton
+            <info.nightscout.androidaps.utils.ui.SingleClickButton
                 android:id="@+id/actions_settempbasal"
                 style="@style/ButtonSmallFontStyle"
                 android:layout_width="0dp"
@@ -70,7 +70,7 @@
                 app:layout_gravity="fill"
                 app:layout_row="1" />
 
-            <info.nightscout.androidaps.utils.SingleClickButton
+            <info.nightscout.androidaps.utils.ui.SingleClickButton
                 android:id="@+id/actions_canceltempbasal"
                 style="@style/ButtonSmallFontStyle"
                 android:layout_width="0dp"
@@ -86,7 +86,7 @@
                 app:layout_gravity="fill"
                 app:layout_row="1" />
 
-            <info.nightscout.androidaps.utils.SingleClickButton
+            <info.nightscout.androidaps.utils.ui.SingleClickButton
                 android:id="@+id/actions_extendedbolus"
                 style="@style/ButtonSmallFontStyle"
                 android:layout_width="0dp"
@@ -101,7 +101,7 @@
                 app:layout_gravity="fill"
                 app:layout_row="1" />
 
-            <info.nightscout.androidaps.utils.SingleClickButton
+            <info.nightscout.androidaps.utils.ui.SingleClickButton
                 android:id="@+id/actions_extendedbolus_cancel"
                 style="@style/ButtonSmallFontStyle"
                 android:layout_width="0dp"
@@ -139,7 +139,7 @@
             android:padding="10dip"
             app:columnCount="2">
 
-            <info.nightscout.androidaps.utils.SingleClickButton
+            <info.nightscout.androidaps.utils.ui.SingleClickButton
                 android:id="@+id/actions_bgcheck"
                 style="@style/ButtonSmallFontStyle"
                 android:layout_width="0dp"
@@ -154,7 +154,7 @@
                 app:layout_gravity="fill"
                 app:layout_row="2" />
 
-            <info.nightscout.androidaps.utils.SingleClickButton
+            <info.nightscout.androidaps.utils.ui.SingleClickButton
                 android:id="@+id/actions_fill"
                 style="@style/ButtonSmallFontStyle"
                 android:layout_width="0dp"
@@ -169,7 +169,7 @@
                 app:layout_gravity="fill"
                 app:layout_row="2" />
 
-            <info.nightscout.androidaps.utils.SingleClickButton
+            <info.nightscout.androidaps.utils.ui.SingleClickButton
                 android:id="@+id/actions_cgmsensorinsert"
                 style="@style/ButtonSmallFontStyle"
                 android:layout_width="0dp"
@@ -183,7 +183,7 @@
                 app:layout_gravity="fill"
                 app:layout_row="3" />
 
-            <info.nightscout.androidaps.utils.SingleClickButton
+            <info.nightscout.androidaps.utils.ui.SingleClickButton
                 android:id="@+id/actions_pumpbatterychange"
                 style="@style/ButtonSmallFontStyle"
                 android:layout_width="0dp"
@@ -197,7 +197,7 @@
                 app:layout_gravity="fill"
                 app:layout_row="3" />
 
-            <info.nightscout.androidaps.utils.SingleClickButton
+            <info.nightscout.androidaps.utils.ui.SingleClickButton
                 android:id="@+id/actions_note"
                 style="@style/ButtonSmallFontStyle"
                 android:layout_width="0dp"
@@ -211,7 +211,7 @@
                 app:layout_gravity="fill"
                 app:layout_row="4" />
 
-            <info.nightscout.androidaps.utils.SingleClickButton
+            <info.nightscout.androidaps.utils.ui.SingleClickButton
                 android:id="@+id/actions_exercise"
                 style="@style/ButtonSmallFontStyle"
                 android:layout_width="0dp"
@@ -225,7 +225,7 @@
                 app:layout_gravity="fill"
                 app:layout_row="4" />
 
-            <info.nightscout.androidaps.utils.SingleClickButton
+            <info.nightscout.androidaps.utils.ui.SingleClickButton
                 android:id="@+id/actions_announcement"
                 style="@style/ButtonSmallFontStyle"
                 android:layout_width="0dp"
@@ -239,7 +239,7 @@
                 app:layout_gravity="fill"
                 app:layout_row="5" />
 
-            <info.nightscout.androidaps.utils.SingleClickButton
+            <info.nightscout.androidaps.utils.ui.SingleClickButton
                 android:id="@+id/actions_question"
                 style="@style/ButtonSmallFontStyle"
                 android:layout_width="0dp"
@@ -268,7 +268,7 @@
             android:padding="10dip"
             app:columnCount="2">
 
-            <info.nightscout.androidaps.utils.SingleClickButton
+            <info.nightscout.androidaps.utils.ui.SingleClickButton
                 android:id="@+id/actions_historybrowser"
                 style="@style/ButtonSmallFontStyle"
                 android:layout_width="0dp"
@@ -282,7 +282,7 @@
                 app:layout_gravity="fill"
                 app:layout_row="6" />
 
-            <info.nightscout.androidaps.utils.SingleClickButton
+            <info.nightscout.androidaps.utils.ui.SingleClickButton
                 android:id="@+id/actions_tddstats"
                 style="@style/ButtonSmallFontStyle"
                 android:layout_width="0dp"
