+++ app/src/main/res/layout/activity_smscommunicator_otp.xml
@@ -86,7 +86,7 @@
                 style="@style/warning_label"
                 android:text="@string/smscommunicator_otp_reset_warning" />
 
-            <info.nightscout.androidaps.utils.SingleClickButton
+            <info.nightscout.androidaps.utils.ui.SingleClickButton
                 android:id="@+id/actions_smscommunicator_otp_reset"
                 style="?android:attr/buttonStyle"
                 android:layout_width="fill_parent"
