+++ app/src/main/res/layout/combopump_fragment.xml
@@ -487,7 +487,7 @@
             android:layout_height="wrap_content"
             android:orientation="horizontal">
 
-            <info.nightscout.androidaps.utils.SingleClickButton
+            <info.nightscout.androidaps.utils.ui.SingleClickButton
                 android:id="@+id/combo_refresh_button"
                 style="@style/ButtonSmallFontStyle"
                 android:layout_width="match_parent"
