+++ app/src/main/res/layout/overview_fragment_nsclient_tablet.xml
@@ -476,7 +476,7 @@
                 android:orientation="horizontal"
                 android:paddingRight="5dp">
 
-                <info.nightscout.androidaps.utils.SingleClickButton
+                <info.nightscout.androidaps.utils.ui.SingleClickButton
                     android:id="@+id/overview_treatmentbutton"
                     style="?android:attr/buttonStyle"
                     android:layout_width="0px"
@@ -490,7 +490,7 @@
                     android:textColor="@color/colorTreatmentButton"
                     android:textSize="10sp" />
 
-                <info.nightscout.androidaps.utils.SingleClickButton
+                <info.nightscout.androidaps.utils.ui.SingleClickButton
                     android:id="@+id/overview_insulinbutton"
                     style="?android:attr/buttonStyle"
                     android:layout_width="0px"
@@ -505,7 +505,7 @@
                     android:textSize="10sp"
                     android:visibility="gone" />
 
-                <info.nightscout.androidaps.utils.SingleClickButton
+                <info.nightscout.androidaps.utils.ui.SingleClickButton
                     android:id="@+id/overview_carbsbutton"
                     style="?android:attr/buttonStyle"
                     android:layout_width="0px"
@@ -520,7 +520,7 @@
                     android:textSize="10sp"
                     android:visibility="gone" />
 
-                <info.nightscout.androidaps.utils.SingleClickButton
+                <info.nightscout.androidaps.utils.ui.SingleClickButton
                     android:id="@+id/overview_wizardbutton"
                     style="?android:attr/buttonStyle"
                     android:layout_width="0px"
@@ -535,7 +535,7 @@
                     android:textSize="10sp"
                     android:visibility="gone" />
 
-                <info.nightscout.androidaps.utils.SingleClickButton
+                <info.nightscout.androidaps.utils.ui.SingleClickButton
                     android:id="@+id/overview_quickwizardbutton"
                     style="?android:attr/buttonStyle"
                     android:layout_width="0px"
