+++ omnipod/src/main/res/layout/omnipod_pod_mgmt.xml
@@ -29,7 +29,7 @@
                 android:gravity="center"
                 android:textStyle="bold" />
 
-            <info.nightscout.androidaps.utils.SingleClickButton
+            <info.nightscout.androidaps.utils.ui.SingleClickButton
                 android:id="@+id/initpod_init_pod"
                 style="?android:attr/buttonStyle"
                 android:layout_width="fill_parent"
@@ -50,7 +50,7 @@
                 android:layout_weight="0.5"
                 android:text="" />
 
-            <info.nightscout.androidaps.utils.SingleClickButton
+            <info.nightscout.androidaps.utils.ui.SingleClickButton
                 android:id="@+id/initpod_remove_pod"
                 style="?android:attr/buttonStyle"
                 android:layout_width="fill_parent"
@@ -71,7 +71,7 @@
                 android:layout_weight="0.5"
                 android:text="" />
 
-            <info.nightscout.androidaps.utils.SingleClickButton
+            <info.nightscout.androidaps.utils.ui.SingleClickButton
                 android:id="@+id/initpod_reset_pod"
                 style="?android:attr/buttonStyle"
                 android:layout_width="fill_parent"
@@ -93,7 +93,7 @@
                 android:text="" />
 
 
-            <info.nightscout.androidaps.utils.SingleClickButton
+            <info.nightscout.androidaps.utils.ui.SingleClickButton
                 android:id="@+id/initpod_pod_history"
                 style="?android:attr/buttonStyle"
                 android:layout_width="fill_parent"
